requirements.txt
pandas
requests
Pillow
reportlab
openpyxl
